% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks_cmpnnts.R
\name{attr_cmpnnt_err}
\alias{attr_cmpnnt_err}
\title{Check select multiple binary or rank position attributes}
\usage{
attr_cmpnnt_err(x)
}
\description{
Check select multiple binary or rank position attributes
}
