% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks.R
\name{attr_err}
\alias{attr_err}
\title{Check attribute errors for attributes appearing in all borg classes}
\usage{
attr_err(relevant, label, constraint)
}
\description{
Check attribute errors for attributes appearing in all borg classes
}
