% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks.R
\name{attr_slct_err}
\alias{attr_slct_err}
\title{Check select question choice names and labels}
\usage{
attr_slct_err(choice_names, choice_labels)
}
\description{
Check select question choice names and labels
}
