% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_calc_character.R
\name{calc_character}
\alias{calc_character}
\title{Calculation character constructor}
\usage{
calc_character(
  x = character(),
  relevant = NA,
  label = NA,
  constraint = NA,
  calculation = NA
)
}
\description{
`calc_character()` constructs a calc character vector.
}
