% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_calc_number.R
\name{calc_number}
\alias{calc_number}
\title{Calc number constructor}
\usage{
calc_number(
  x = double(),
  relevant = NA,
  label = NA,
  constraint = NA,
  calculation = NA
)
}
\description{
`calc_number()` constructs a calc_number vector.
}
