% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_calendar.R
\name{calendar}
\alias{calendar}
\title{Calendar constructor}
\usage{
calendar(
  x = new_date(),
  relevant = NA,
  label = NA,
  constraint = NA,
  source = "excel"
)
}
\description{
`calendar()` constructs a calendar vector.
}
