% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_calc_number.R
\name{cast-borg-calc-number}
\alias{cast-borg-calc-number}
\alias{as_calc_number}
\alias{as_calc_number.borg_calc_number}
\alias{as_calc_number.integer}
\alias{as_calc_number.double}
\alias{as_calc_number.character}
\title{Cast to `calc_number`}
\usage{
as_calc_number(x, ...)

\method{as_calc_number}{borg_calc_number}(x, ...)

\method{as_calc_number}{integer}(
  x,
  relevant = NA,
  label = NA,
  constraint = NA,
  calculation = NA,
  ...
)

\method{as_calc_number}{double}(
  x,
  relevant = NA,
  label = NA,
  constraint = NA,
  calculation = NA,
  ...
)

\method{as_calc_number}{character}(
  x,
  relevant = NA,
  label = NA,
  constraint = NA,
  calculation = NA,
  ...
)
}
\arguments{
\item{x}{An object to coerce to `calc_number`.}

\item{...}{Arguments passed on to further methods.}
}
\description{
Cast `x` to a `calc_number` vector
}
