% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_number.R
\name{cast-borg-number}
\alias{cast-borg-number}
\alias{as_number}
\alias{as_number.borg_number}
\alias{as_number.integer}
\alias{as_number.double}
\alias{as_number.character}
\title{Cast to `number`}
\usage{
as_number(x, ...)

\method{as_number}{borg_number}(x, ...)

\method{as_number}{integer}(x, relevant = NA, label = NA, constraint = NA, ...)

\method{as_number}{double}(x, relevant = NA, label = NA, constraint = NA, ...)

\method{as_number}{character}(x, relevant = NA, label = NA, constraint = NA, ...)
}
\arguments{
\item{x}{An object to coerce to `number`.}

\item{...}{Arguments passed on to further methods.}
}
\description{
Cast `x` to a `number` vector
}
