% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_calc_character.R
\name{cast-calc-character}
\alias{cast-calc-character}
\alias{as_calc_character}
\alias{as_calc_character.borg_calc_character}
\alias{as_calc_character.character}
\title{Cast to `borg_calc_character`}
\usage{
as_calc_character(x, ...)

\method{as_calc_character}{borg_calc_character}(x, ...)

\method{as_calc_character}{character}(
  x,
  relevant = NA,
  label = NA,
  constraint = NA,
  calculation = NA,
  ...
)
}
\arguments{
\item{x}{An object to coerce to `borg_calc_character`.}

\item{...}{Arguments passed on to further methods.}
}
\description{
Cast `x` to a `borg_calc_character` vector
}
