% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_calendar.R
\name{cast-calendar}
\alias{cast-calendar}
\alias{as_calendar}
\alias{as_calendar.borg_calendar}
\alias{as_calendar.character}
\title{Cast to `borg_calendar`}
\usage{
as_calendar(x, ...)

\method{as_calendar}{borg_calendar}(x, ...)

\method{as_calendar}{character}(x, relevant = NA, label = NA, constraint = NA, ...)
}
\arguments{
\item{x}{An object to coerce to `borg_calendar`.}

\item{...}{Arguments passed on to further methods.}
}
\description{
Cast `x` to a `borg_calendar` vector
}
