% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_rank.R
\name{cast-rank-char}
\alias{cast-rank-char}
\alias{as_rank_char}
\alias{as_rank_char.borg_rank_char}
\alias{as_rank_char.character}
\alias{as_rank_char.borg_rank_list}
\alias{as_rank_list.character}
\alias{as_rank_list.borg_rank_char}
\title{Cast to `borg_rank_char`}
\usage{
as_rank_char(x, ...)

\method{as_rank_char}{borg_rank_char}(x, ...)

\method{as_rank_char}{character}(
  x,
  relevant = NA,
  choice_names = NA,
  choice_labels = NA,
  q_name = NA,
  label = NA,
  constraint = NA,
  max_rank = na_length(choice_names),
  position_sep = "/",
  ...
)

\method{as_rank_char}{borg_rank_list}(x, ...)

\method{as_rank_list}{character}(
  x,
  relevant = NA,
  choice_names = NA,
  choice_labels = NA,
  q_name = NA,
  label = NA,
  constraint = NA,
  max_rank = na_length(choice_names),
  position_sep = "/",
  ...
)

\method{as_rank_list}{borg_rank_char}(x, ...)
}
\arguments{
\item{x}{An object to coerce to `borg_rank_char`.}

\item{...}{Arguments passed on to further methods.}
}
\description{
Cast `x` to a `borg_rank_char` vector
}
