% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_rank.R
\name{cast-rank-list}
\alias{cast-rank-list}
\alias{as_rank_list}
\alias{as_rank_list.borg_rank_list}
\title{Cast to `borg_rank_list`}
\usage{
as_rank_list(x, ...)

\method{as_rank_list}{borg_rank_list}(x, ...)
}
\arguments{
\item{x}{An object to coerce to `borg_rank_list`.}

\item{...}{Arguments passed on to further methods.}
}
\description{
Cast `x` to a `borg_rank_list` vector
}
