% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_rank_position.R
\name{cast-rank-position}
\alias{cast-rank-position}
\alias{as_rank_position}
\alias{as_rank_position.borg_rank_position}
\alias{as_rank_position.numeric}
\alias{as_rank_position.double}
\alias{as_rank_position.integer}
\title{Cast to `borg_rank_position`}
\usage{
as_rank_position(x, ...)

\method{as_rank_position}{borg_rank_position}(x, ...)

\method{as_rank_position}{numeric}(
  x,
  relevant = NA,
  choice_name = NA,
  choice_label = NA,
  q_name = NA,
  label = NA,
  constraint = NA,
  max_rank = NA,
  position_sep = "/",
  ...
)

\method{as_rank_position}{double}(
  x,
  relevant = NA,
  choice_name = NA,
  choice_label = NA,
  q_name = NA,
  label = NA,
  constraint = NA,
  max_rank = NA,
  position_sep = "/",
  ...
)

\method{as_rank_position}{integer}(
  x,
  relevant = NA,
  choice_name = NA,
  choice_label = NA,
  q_name = NA,
  label = NA,
  constraint = NA,
  max_rank = NA,
  position_sep = "/",
  ...
)
}
\arguments{
\item{x}{An object to coerce to `borg_rank_position`.}

\item{...}{Arguments passed on to further methods.}
}
\description{
Cast `x` to a `borg_rank_position` vector
}
