% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_select_one.R
\name{cast-slct-one}
\alias{cast-slct-one}
\alias{as_slct_one}
\alias{as_slct_one.borg_slct_one}
\alias{as_slct_one.character}
\title{Cast to `borg_slct_one`}
\usage{
as_slct_one(x, ...)

\method{as_slct_one}{borg_slct_one}(x, ...)

\method{as_slct_one}{character}(
  x,
  relevant = NA,
  choice_names = NA,
  choice_labels = NA,
  label = NA,
  constraint = NA,
  ...
)
}
\arguments{
\item{x}{An object to coerce to `borg_slct_one`.}

\item{...}{Arguments passed on to further methods.}
}
\description{
Cast `x` to a `borg_slct_one` vector
}
