% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_select_multiple_binary.R
\name{cast-sm-binary}
\alias{cast-sm-binary}
\alias{as_sm_binary}
\alias{as_sm_binary.borg_sm_binary}
\alias{as_sm_binary.logical}
\alias{as_sm_binary.numeric}
\alias{as_sm_binary.double}
\alias{as_sm_binary.integer}
\title{Cast to `borg_sm_binary`}
\usage{
as_sm_binary(x, ...)

\method{as_sm_binary}{borg_sm_binary}(x, ...)

\method{as_sm_binary}{logical}(
  x,
  relevant = NA,
  choice_name = NA,
  choice_label = NA,
  q_name = NA,
  label = NA,
  constraint = NA,
  binary_sep = "/",
  ...
)

\method{as_sm_binary}{numeric}(
  x,
  relevant = NA,
  choice_name = NA,
  choice_label = NA,
  q_name = NA,
  label = NA,
  constraint = NA,
  binary_sep = "/",
  ...
)

\method{as_sm_binary}{double}(
  x,
  relevant = NA,
  choice_name = NA,
  choice_label = NA,
  q_name = NA,
  label = NA,
  constraint = NA,
  binary_sep = "/",
  ...
)

\method{as_sm_binary}{integer}(
  x,
  relevant = NA,
  choice_name = NA,
  choice_label = NA,
  q_name = NA,
  label = NA,
  constraint = NA,
  binary_sep = "/",
  ...
)
}
\arguments{
\item{x}{An object to coerce to `borg_sm_binary`.}

\item{...}{Arguments passed on to further methods.}
}
\description{
Cast `x` to a `borg_sm_binary` vector
}
