% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_select_multiple.R
\name{cast-sm-char}
\alias{cast-sm-char}
\alias{as_sm_char}
\alias{as_sm_char.borg_sm_char}
\alias{as_sm_char.character}
\alias{as_sm_char.borg_sm_list}
\alias{as_sm_list.character}
\alias{as_sm_list.borg_sm_char}
\title{Cast to `borg_sm_char`}
\usage{
as_sm_char(x, ...)

\method{as_sm_char}{borg_sm_char}(x, ...)

\method{as_sm_char}{character}(
  x,
  relevant = NA,
  choice_names = NA,
  choice_labels = NA,
  label = NA,
  constraint = NA,
  binary_sep = "/",
  ...
)

\method{as_sm_char}{borg_sm_list}(x, ...)

\method{as_sm_list}{character}(
  x,
  relevant = NA,
  choice_names = NA,
  choice_labels = NA,
  label = NA,
  constraint = NA,
  binary_sep = "/",
  ...
)

\method{as_sm_list}{borg_sm_char}(x, ...)
}
\arguments{
\item{x}{An object to coerce to `borg_sm_char`.}

\item{...}{Arguments passed on to further methods.}
}
\description{
Cast `x` to a `borg_sm_char` vector
}
