% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_select_multiple.R
\name{cast-sm-list}
\alias{cast-sm-list}
\alias{as_sm_list}
\alias{as_sm_list.borg_sm_list}
\title{Cast to `borg_sm_list`}
\usage{
as_sm_list(x, ...)

\method{as_sm_list}{borg_sm_list}(x, ...)
}
\arguments{
\item{x}{An object to coerce to `borg_sm_list`.}

\item{...}{Arguments passed on to further methods.}
}
\description{
Cast `x` to a `borg_sm_list` vector
}
