% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_text.R
\name{cast-text}
\alias{cast-text}
\alias{as_text}
\alias{as_text.borg_text}
\alias{as_text.character}
\title{Cast to `borg_text`}
\usage{
as_text(x, ...)

\method{as_text}{borg_text}(x, ...)

\method{as_text}{character}(x, relevant = NA, label = NA, constraint = NA, ...)
}
\arguments{
\item{x}{An object to coerce to `borg_text`.}

\item{...}{Arguments passed on to further methods.}
}
\description{
Cast `x` to a `borg_text` vector
}
