% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks.R
\name{identical_borg_attr}
\alias{identical_borg_attr}
\title{Check if attributes identical for generic borg questions}
\usage{
identical_borg_attr(x, y)
}
\description{
Check if attributes identical for generic borg questions
}
