% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks.R
\name{identical_calc_attr}
\alias{identical_calc_attr}
\title{Check if attributes identical for calculation borg questions}
\usage{
identical_calc_attr(x, y)
}
\description{
Check if attributes identical for calculation borg questions
}
