% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks_rnk.R
\name{identical_rnk_attr}
\alias{identical_rnk_attr}
\title{Check if attributes identical for rank questions}
\usage{
identical_rnk_attr(x, y)
}
\description{
Check if attributes identical for rank questions
}
