% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks.R
\name{identical_sm_attr}
\alias{identical_sm_attr}
\title{Check if attributes identical for select multiple questions}
\usage{
identical_sm_attr(x, y)
}
\description{
Check if attributes identical for select multiple questions
}
