% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks_so.R
\name{identical_so_attr}
\alias{identical_so_attr}
\title{Check if attributes identical for select one questions}
\usage{
identical_so_attr(x, y)
}
\description{
Check if attributes identical for select one questions
}
