% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_calc_character.R
\name{new_calc_character}
\alias{new_calc_character}
\title{Low level calc character constructor}
\usage{
new_calc_character(
  x = character(),
  relevant = NA,
  label = NA,
  constraint = NA,
  calculation = NA
)
}
\description{
Low level calc character constructor
}
