% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_calc_number.R
\name{new_calc_number}
\alias{new_calc_number}
\title{Low level calc number constructor}
\usage{
new_calc_number(
  x = double(),
  relevant = NA,
  label = NA,
  constraint = NA,
  calculation = NA
)
}
\description{
Low level calc number constructor
}
