% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_calendar.R
\name{new_calendar}
\alias{new_calendar}
\title{Low level calendar constructor}
\usage{
new_calendar(x = new_date(), relevant = NA, label = NA, constraint = NA)
}
\description{
Low level calendar constructor
}
