% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_rank.R
\name{new_rank_char}
\alias{new_rank_char}
\title{Low level rank character constructor}
\usage{
new_rank_char(
  x = character(),
  relevant = NA,
  choice_names = NA,
  choice_labels = NA,
  q_name = NA,
  label = NA,
  constraint = NA,
  binary_sep = "/"
)
}
\description{
Low level rank character constructor
}
