% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_rank.R
\name{new_rank_list}
\alias{new_rank_list}
\title{Low level rank list constructor}
\usage{
new_rank_list(
  x = list(),
  relevant = NA,
  choice_names = NA,
  choice_labels = NA,
  q_name = NA,
  label = NA,
  constraint = NA,
  max_rank = na_length(choice_names),
  position_sep = "/"
)
}
\description{
Low level rank list constructor
}
