% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_rank_position.R
\name{new_rank_position}
\alias{new_rank_position}
\title{Low level rank position constructor}
\usage{
new_rank_position(
  x = integer(),
  relevant = NA,
  choice_name = NA,
  choice_label = NA,
  q_name = NA,
  label = NA,
  constraint = NA,
  max_rank = NA,
  position_sep = "/"
)
}
\description{
Low level rank position constructor
}
