% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_select_one.R
\name{new_slct_one}
\alias{new_slct_one}
\title{Low level select one constructor}
\usage{
new_slct_one(
  x = character(),
  relevant = NA,
  choice_names = NA,
  choice_labels = NA,
  label = NA,
  constraint = NA
)
}
\description{
Low level select one constructor
}
