% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_select_multiple_binary.R
\name{new_sm_binary}
\alias{new_sm_binary}
\title{Low level select multiple binary constructor}
\usage{
new_sm_binary(
  x = logical(),
  relevant = NA,
  choice_name = NA,
  choice_label = NA,
  q_name = NA,
  label = NA,
  constraint = NA,
  binary_sep = "/"
)
}
\description{
Low level select multiple binary constructor
}
