% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_select_multiple.R
\name{new_sm_list}
\alias{new_sm_list}
\title{Low level select multiple list constructor}
\usage{
new_sm_list(
  x = list(),
  relevant = NA,
  choice_names = NA,
  choice_labels = NA,
  label = NA,
  constraint = NA,
  binary_sep = "/"
)
}
\description{
Low level select multiple list constructor
}
