% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_rank.R
\name{rank_char}
\alias{rank_char}
\title{Rank character constructor}
\usage{
rank_char(
  x = character(),
  relevant = NA,
  choice_names = NA,
  choice_labels = NA,
  q_name = NA,
  label = NA,
  constraint = NA,
  max_rank = na_length(choice_names),
  position_sep = "/"
)
}
\description{
`rank_char()` constructs a rank character vector, each value in the vector corresponds to the ordered
options for that survey row in string format. Can be constructed from a character vector itself.
}
