% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_rank_position.R
\name{rank_position}
\alias{rank_position}
\title{Rank position constructor}
\usage{
rank_position(
  x = integer(),
  relevant = NA,
  choice_name = NA,
  choice_label = NA,
  q_name = NA,
  label = NA,
  constraint = NA,
  position_sep = "/"
)
}
\description{
`rank_position()` constructs a rank position vector, which is an integer vector where each row
indicates the rank of that particular choice option. This output is not provided by KoBo outputs
by default.
}
