% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks_rnk.R
\name{rnk_max_rank_error}
\alias{rnk_max_rank_error}
\title{Check max rank attribute for borg position classes}
\usage{
rnk_max_rank_error(x)
}
\description{
Check max rank attribute for borg position classes
}
