% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks_rnk.R
\name{rnk_pos_rank_check}
\alias{rnk_pos_rank_check}
\title{Check that position vector doesn't taken on value higher than max_rank}
\usage{
rnk_pos_rank_check(x, max_rank)
}
\description{
Check that position vector doesn't taken on value higher than max_rank
}
