% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks_rnk.R
\name{rnk_pos_sep_error}
\alias{rnk_pos_sep_error}
\title{Check position separator attribute for borg rank position classes}
\usage{
rnk_pos_sep_error(x)
}
\description{
Check position separator attribute for borg rank position classes
}
