% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_select_one.R
\name{slct_one}
\alias{slct_one}
\title{Select one character constructor}
\usage{
slct_one(
  x = character(),
  relevant = NA,
  choice_names = NA,
  choice_labels = NA,
  label = NA,
  constraint = NA
)
}
\description{
`slct_one()` constructs a select one vector. Can be constructed from a character vector itself.
}
