% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks.R
\name{sm_bin_sep_error}
\alias{sm_bin_sep_error}
\title{Check binary separator attribute for borg select multiple classes}
\usage{
sm_bin_sep_error(x)
}
\description{
Check binary separator attribute for borg select multiple classes
}
