% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_select_multiple_binary.R
\name{sm_binary}
\alias{sm_binary}
\title{Select multiple binary constructor}
\usage{
sm_binary(
  x = character(),
  relevant = NA,
  choice_name = NA,
  choice_label = NA,
  q_name = NA,
  label = NA,
  constraint = NA,
  binary_type = "logical",
  binary_sep = "/"
)
}
\description{
`sm_binary()` constructs a select multiple binary vector, which is a logical vector where each row
indicates if that particular choice option was selected or not.
}
