% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_select_multiple.R
\name{sm_char}
\alias{sm_char}
\title{Select multiple character constructor}
\usage{
sm_char(
  x = character(),
  relevant = NA,
  choice_names = NA,
  choice_labels = NA,
  label = NA,
  constraint = NA,
  binary_type = "logical",
  binary_sep = "/"
)
}
\description{
`sm_char()` constructs a select multiple character vector, each value in the vecotr corresponds to the selected
options for that survey row in string format. Can be constructed from a character vector itself.
}
