% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks_sm.R
\name{sm_chc_check}
\alias{sm_chc_check}
\title{Check if select multiple values in choices}
\usage{
sm_chc_check(x, choice_names, type = "list")
}
\description{
Check if select multiple values in choices
}
