% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_select_multiple.R
\name{sm_list}
\alias{sm_list}
\title{Select multiple list constructor}
\usage{
sm_list(
  x = list(),
  relevant = NA,
  choice_names = NA,
  choice_labels = NA,
  label = NA,
  constraint = NA,
  binary_sep = "/"
)
}
\description{
`sm_list()` constructs a select multiple list object, where each entry in the list corresponds to the selected
options for that survey row. Can be constructed from an existing list of character vectors or a character vector
itself that will be turned into a list by splitting on spaces in the string.
}
