% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_rank_position.R
\name{validate_rank_pos}
\alias{validate_rank_pos}
\title{Validation function for both select multiple classes}
\usage{
validate_rank_pos(x)
}
\description{
Validation function for both select multiple classes
}
