% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_select_multiple.R,
%   R/class_select_multiple_binary.R
\name{validate_sm}
\alias{validate_sm}
\title{Validation function for both select multiple classes}
\usage{
validate_sm(x)

validate_sm(x)
}
\description{
Validation function for both select multiple classes

Validation function for both select multiple classes
}
