% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_calc_character.R
\name{vec_cast.borg_calc_character.character}
\alias{vec_cast.borg_calc_character.character}
\title{Casting character to borg_calc_characteracter}
\usage{
\method{vec_cast.borg_calc_character}{character}(x, to, ...)
}
\description{
Casting character to borg_calc_characteracter
}
