% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_calc_character.R
\name{vec_cast.borg_calc_character}
\alias{vec_cast.borg_calc_character}
\title{Boiler plate for casting calc character}
\usage{
\method{vec_cast}{borg_calc_character}(x, to, ...)
}
\description{
Boiler plate for casting calc character
}
