% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_calc_number.R
\name{vec_cast.borg_calc_number.borg_calc_number}
\alias{vec_cast.borg_calc_number.borg_calc_number}
\title{Casting calc number to calc number}
\usage{
\method{vec_cast.borg_calc_number}{borg_calc_number}(x, to, ...)
}
\description{
Casting calc number to calc number
}
