% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_calc_number.R
\name{vec_cast.borg_calc_number.double}
\alias{vec_cast.borg_calc_number.double}
\title{Casting double to calc number}
\usage{
\method{vec_cast.borg_calc_number}{double}(x, to, ...)
}
\description{
Casting double to calc number
}
