% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_calc_number.R
\name{vec_cast.borg_calc_number.integer}
\alias{vec_cast.borg_calc_number.integer}
\title{Casting integer to calc number}
\usage{
\method{vec_cast.borg_calc_number}{integer}(x, to, ...)
}
\description{
Casting integer to calc number
}
