% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_calendar.R
\name{vec_cast.borg_calendar.borg_calendar}
\alias{vec_cast.borg_calendar.borg_calendar}
\title{Casting borg_calendar to borg_calendar}
\usage{
\method{vec_cast.borg_calendar}{borg_calendar}(x, to, ...)
}
\description{
Casting borg_calendar to borg_calendar
}
