% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_calendar.R
\name{vec_cast.borg_calendar.character}
\alias{vec_cast.borg_calendar.character}
\title{Casting character to borg_calendaracter}
\usage{
\method{vec_cast.borg_calendar}{character}(x, to, ...)
}
\description{
Casting character to borg_calendaracter
}
