% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_calendar.R
\name{vec_cast.borg_calendar.Date}
\alias{vec_cast.borg_calendar.Date}
\title{Casting Date to borg_calendaracter}
\usage{
\method{vec_cast.borg_calendar}{Date}(x, to, ...)
}
\description{
Casting Date to borg_calendaracter
}
