% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_calendar.R
\name{vec_cast.borg_calendar.POSIXct}
\alias{vec_cast.borg_calendar.POSIXct}
\title{Casting POSIXct to borg_calendaracter}
\usage{
\method{vec_cast.borg_calendar}{POSIXct}(x, to, ...)
}
\description{
Casting POSIXct to borg_calendaracter
}
