% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_calendar.R
\name{vec_cast.borg_calendar.POSIXlt}
\alias{vec_cast.borg_calendar.POSIXlt}
\title{Casting POSIXlt to borg_calendaracter}
\usage{
\method{vec_cast.borg_calendar}{POSIXlt}(x, to, ...)
}
\description{
Casting POSIXlt to borg_calendaracter
}
