% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_calendar.R
\name{vec_cast.borg_calendar}
\alias{vec_cast.borg_calendar}
\title{Boiler plate for casting calendar}
\usage{
\method{vec_cast}{borg_calendar}(x, to, ...)
}
\description{
Boiler plate for casting calendar
}
