% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_number.R
\name{vec_cast.borg_number.borg_number}
\alias{vec_cast.borg_number.borg_number}
\title{Casting number to number}
\usage{
\method{vec_cast.borg_number}{borg_number}(x, to, ...)
}
\description{
Casting number to number
}
