% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_number.R
\name{vec_cast.borg_number.character}
\alias{vec_cast.borg_number.character}
\title{Casting character to number}
\usage{
\method{vec_cast.borg_number}{character}(x, to, ...)
}
\description{
Casting character to number
}
