% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_number.R
\name{vec_cast.borg_number.integer}
\alias{vec_cast.borg_number.integer}
\title{Casting integer to number}
\usage{
\method{vec_cast.borg_number}{integer}(x, to, ...)
}
\description{
Casting integer to number
}
