% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_rank.R
\name{vec_cast.borg_rank_char.borg_rank_list}
\alias{vec_cast.borg_rank_char.borg_rank_list}
\title{Casting borg_rank_list to borg_rank_char}
\usage{
\method{vec_cast.borg_rank_char}{borg_rank_list}(x, to, ...)
}
\description{
Casting borg_rank_list to borg_rank_char
}
