% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_rank.R
\name{vec_cast.borg_rank_char.character}
\alias{vec_cast.borg_rank_char.character}
\title{Casting character to borg_rank_character}
\usage{
\method{vec_cast.borg_rank_char}{character}(x, to, ...)
}
\description{
Casting character to borg_rank_character
}
