% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_rank.R
\name{vec_cast.borg_rank_char}
\alias{vec_cast.borg_rank_char}
\title{Boiler plate for casting rank character}
\usage{
\method{vec_cast}{borg_rank_char}(x, to, ...)
}
\description{
Boiler plate for casting rank character
}
