% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_rank.R
\name{vec_cast.borg_rank_list.borg_rank_char}
\alias{vec_cast.borg_rank_list.borg_rank_char}
\title{Casting borg_rank_char to borg_rank_list}
\usage{
\method{vec_cast.borg_rank_list}{borg_rank_char}(x, to, ...)
}
\description{
Casting borg_rank_char to borg_rank_list
}
