% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_rank.R
\name{vec_cast.borg_rank_list}
\alias{vec_cast.borg_rank_list}
\title{Boiler plate for casting rank list}
\usage{
\method{vec_cast}{borg_rank_list}(x, to, ...)
}
\description{
Boiler plate for casting rank list
}
