% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_rank_position.R
\name{vec_cast.borg_rank_position.borg_rank_position}
\alias{vec_cast.borg_rank_position.borg_rank_position}
\title{Casting borg_rank_position to borg_rank_position}
\usage{
\method{vec_cast.borg_rank_position}{borg_rank_position}(x, to, ...)
}
\description{
Casting borg_rank_position to borg_rank_position
}
