% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_rank_position.R
\name{vec_cast.borg_rank_position.double}
\alias{vec_cast.borg_rank_position.double}
\title{Casting double to borg_rank_positionacter}
\usage{
\method{vec_cast.borg_rank_position}{double}(x, to, ...)
}
\description{
Casting double to borg_rank_positionacter
}
