% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_rank_position.R
\name{vec_cast.borg_rank_position.integer}
\alias{vec_cast.borg_rank_position.integer}
\title{Casting integer to borg_rank_positionacter}
\usage{
\method{vec_cast.borg_rank_position}{integer}(x, to, ...)
}
\description{
Casting integer to borg_rank_positionacter
}
