% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_rank_position.R
\name{vec_cast.borg_rank_position.logical}
\alias{vec_cast.borg_rank_position.logical}
\title{Casting logical to borg_rank_positionacter}
\usage{
\method{vec_cast.borg_rank_position}{logical}(x, to, ...)
}
\description{
Casting logical to borg_rank_positionacter
}
