% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_rank_position.R
\name{vec_cast.borg_rank_position}
\alias{vec_cast.borg_rank_position}
\title{Boiler plate for casting rank position}
\usage{
\method{vec_cast}{borg_rank_position}(x, to, ...)
}
\description{
Boiler plate for casting rank position
}
