% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_select_one.R
\name{vec_cast.borg_slct_one.borg_slct_one}
\alias{vec_cast.borg_slct_one.borg_slct_one}
\title{Casting borg_slct_one to borg_slct_one}
\usage{
\method{vec_cast.borg_slct_one}{borg_slct_one}(x, to, ...)
}
\description{
Casting borg_slct_one to borg_slct_one
}
