% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_select_one.R
\name{vec_cast.borg_slct_one.character}
\alias{vec_cast.borg_slct_one.character}
\title{Casting character to borg_slct_oneacter}
\usage{
\method{vec_cast.borg_slct_one}{character}(x, to, ...)
}
\description{
Casting character to borg_slct_oneacter
}
