% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_select_multiple_binary.R
\name{vec_cast.borg_sm_binary.double}
\alias{vec_cast.borg_sm_binary.double}
\title{Casting double to borg_sm_binaryacter}
\usage{
\method{vec_cast.borg_sm_binary}{double}(x, to, ...)
}
\description{
Casting double to borg_sm_binaryacter
}
