% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_select_multiple_binary.R
\name{vec_cast.borg_sm_binary.integer}
\alias{vec_cast.borg_sm_binary.integer}
\title{Casting integer to borg_sm_binaryacter}
\usage{
\method{vec_cast.borg_sm_binary}{integer}(x, to, ...)
}
\description{
Casting integer to borg_sm_binaryacter
}
