% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_select_multiple_binary.R
\name{vec_cast.borg_sm_binary.logical}
\alias{vec_cast.borg_sm_binary.logical}
\title{Casting logical to borg_sm_binaryacter}
\usage{
\method{vec_cast.borg_sm_binary}{logical}(x, to, ...)
}
\description{
Casting logical to borg_sm_binaryacter
}
