% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_select_multiple_binary.R
\name{vec_cast.borg_sm_binary}
\alias{vec_cast.borg_sm_binary}
\title{Boiler plate for casting select multiple binary}
\usage{
\method{vec_cast}{borg_sm_binary}(x, to, ...)
}
\description{
Boiler plate for casting select multiple binary
}
