% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_select_multiple.R
\name{vec_cast.borg_sm_char.borg_sm_list}
\alias{vec_cast.borg_sm_char.borg_sm_list}
\title{Casting borg_sm_list to borg_sm_char}
\usage{
\method{vec_cast.borg_sm_char}{borg_sm_list}(x, to, ...)
}
\description{
Casting borg_sm_list to borg_sm_char
}
