% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_select_multiple.R
\name{vec_cast.borg_sm_char.character}
\alias{vec_cast.borg_sm_char.character}
\title{Casting character to borg_sm_character}
\usage{
\method{vec_cast.borg_sm_char}{character}(x, to, ...)
}
\description{
Casting character to borg_sm_character
}
