% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_select_multiple.R
\name{vec_cast.borg_sm_char}
\alias{vec_cast.borg_sm_char}
\title{Boiler plate for casting select multiple character}
\usage{
\method{vec_cast}{borg_sm_char}(x, to, ...)
}
\description{
Boiler plate for casting select multiple character
}
