% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_select_multiple.R
\name{vec_cast.borg_sm_list}
\alias{vec_cast.borg_sm_list}
\title{Boiler plate for casting select multiple list}
\usage{
\method{vec_cast}{borg_sm_list}(x, to, ...)
}
\description{
Boiler plate for casting select multiple list
}
