% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_text.R
\name{vec_cast.borg_text.character}
\alias{vec_cast.borg_text.character}
\title{Casting character to borg_textacter}
\usage{
\method{vec_cast.borg_text}{character}(x, to, ...)
}
\description{
Casting character to borg_textacter
}
