% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_select_multiple_helper.R
\name{vec_cast.char_helper.char_helper}
\alias{vec_cast.char_helper.char_helper}
\title{Casting char_helper to char_helper}
\usage{
\method{vec_cast.char_helper}{char_helper}(x, to, ...)
}
\description{
Casting char_helper to char_helper
}
