% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_select_multiple_helper.R
\name{vec_cast.char_helper.character}
\alias{vec_cast.char_helper.character}
\title{Casting character to char_helperacter}
\usage{
\method{vec_cast.char_helper}{character}(x, to, ...)
}
\description{
Casting character to char_helperacter
}
