% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_select_multiple_helper.R
\name{vec_cast.char_helper}
\alias{vec_cast.char_helper}
\title{Boiler plate for casting text}
\usage{
\method{vec_cast}{char_helper}(x, to, ...)
}
\description{
Boiler plate for casting text
}
