% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_calc_character.R
\name{vec_cast.character.borg_calc_character}
\alias{vec_cast.character.borg_calc_character}
\title{Casting borg_calc_character to character}
\usage{
\method{vec_cast.character}{borg_calc_character}(x, to, ...)
}
\description{
Casting borg_calc_character to character
}
