% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_number.R
\name{vec_cast.character.borg_number}
\alias{vec_cast.character.borg_number}
\title{Casting number to character}
\usage{
\method{vec_cast.character}{borg_number}(x, to, ...)
}
\description{
Casting number to character
}
