% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_rank.R
\name{vec_cast.character.borg_rank_char}
\alias{vec_cast.character.borg_rank_char}
\title{Casting borg_rank_char to character}
\usage{
\method{vec_cast.character}{borg_rank_char}(x, to, ...)
}
\description{
Casting borg_rank_char to character
}
