% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_select_one.R
\name{vec_cast.character.borg_slct_one}
\alias{vec_cast.character.borg_slct_one}
\title{Casting borg_slct_one to character}
\usage{
\method{vec_cast.character}{borg_slct_one}(x, to, ...)
}
\description{
Casting borg_slct_one to character
}
