% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_select_multiple.R
\name{vec_cast.character.borg_sm_char}
\alias{vec_cast.character.borg_sm_char}
\title{Casting borg_sm_char to character}
\usage{
\method{vec_cast.character}{borg_sm_char}(x, to, ...)
}
\description{
Casting borg_sm_char to character
}
