% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_text.R
\name{vec_cast.character.borg_text}
\alias{vec_cast.character.borg_text}
\title{Casting borg_text to character}
\usage{
\method{vec_cast.character}{borg_text}(x, to, ...)
}
\description{
Casting borg_text to character
}
