% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_select_multiple_helper.R
\name{vec_cast.character.char_helper}
\alias{vec_cast.character.char_helper}
\title{Casting char_helper to character}
\usage{
\method{vec_cast.character}{char_helper}(x, to, ...)
}
\description{
Casting char_helper to character
}
