% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_number.R
\name{vec_cast.integer.borg_number}
\alias{vec_cast.integer.borg_number}
\title{Casting number to integer}
\usage{
\method{vec_cast.integer}{borg_number}(x, to, ...)
}
\description{
Casting number to integer
}
