% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_select_multiple_binary.R
\name{vec_cast.integer.borg_sm_binary}
\alias{vec_cast.integer.borg_sm_binary}
\title{Casting borg_sm_binary to integer}
\usage{
\method{vec_cast.integer}{borg_sm_binary}(x, to, ...)
}
\description{
Casting borg_sm_binary to integer
}
