% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_calendar.R
\name{vec_cast.POSIXct.borg_calendar}
\alias{vec_cast.POSIXct.borg_calendar}
\title{Casting borg_calendar to POSIXct}
\usage{
\method{vec_cast.POSIXct}{borg_calendar}(x, to, ...)
}
\description{
Casting borg_calendar to POSIXct
}
