% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_calendar.R
\name{vec_cast.POSIXlt.borg_calendar}
\alias{vec_cast.POSIXlt.borg_calendar}
\title{Casting borg_calendar to POSIXlt}
\usage{
\method{vec_cast.POSIXlt}{borg_calendar}(x, to, ...)
}
\description{
Casting borg_calendar to POSIXlt
}
