% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_calc_character.R
\name{vec_ptype2.borg_calc_character}
\alias{vec_ptype2.borg_calc_character}
\title{Boiler plate for coercion for coercion of calc characters}
\usage{
\method{vec_ptype2}{borg_calc_character}(x, y, ...)
}
\description{
Boiler plate for coercion for coercion of calc characters
}
