% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_calc_number.R
\name{vec_ptype2.borg_calc_number}
\alias{vec_ptype2.borg_calc_number}
\title{Boiler plate for coercion for coercion of calc number}
\usage{
\method{vec_ptype2}{borg_calc_number}(x, y, ...)
}
\description{
Boiler plate for coercion for coercion of calc number
}
