% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_calendar.R
\name{vec_ptype2.borg_calendar}
\alias{vec_ptype2.borg_calendar}
\title{Boiler plate for coercion for coercion of calendars}
\usage{
\method{vec_ptype2}{borg_calendar}(x, y, ...)
}
\description{
Boiler plate for coercion for coercion of calendars
}
