% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_number.R
\name{vec_ptype2.borg_number}
\alias{vec_ptype2.borg_number}
\title{Boiler plate for coercion for coercion of select one characters}
\usage{
\method{vec_ptype2}{borg_number}(x, y, ...)
}
\description{
Boiler plate for coercion for coercion of select one characters
}
