% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_rank.R
\name{vec_ptype2.borg_rank_char}
\alias{vec_ptype2.borg_rank_char}
\title{Boiler plate for coercion for coercion of rank characters}
\usage{
\method{vec_ptype2}{borg_rank_char}(x, y, ...)
}
\description{
Boiler plate for coercion for coercion of rank characters
}
