% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_rank.R
\name{vec_ptype2.borg_rank_list}
\alias{vec_ptype2.borg_rank_list}
\title{Boiler plate for coercion of rank lists}
\usage{
\method{vec_ptype2}{borg_rank_list}(x, y, ...)
}
\description{
Boiler plate for coercion of rank lists
}
