% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_rank_position.R
\name{vec_ptype2.borg_rank_position}
\alias{vec_ptype2.borg_rank_position}
\title{Boiler plate for coercion for coercion of rank positions}
\usage{
\method{vec_ptype2}{borg_rank_position}(x, y, ...)
}
\description{
Boiler plate for coercion for coercion of rank positions
}
