% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_select_one.R
\name{vec_ptype2.borg_slct_one}
\alias{vec_ptype2.borg_slct_one}
\title{Boiler plate for coercion for coercion of select one characters}
\usage{
\method{vec_ptype2}{borg_slct_one}(x, y, ...)
}
\description{
Boiler plate for coercion for coercion of select one characters
}
