% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_select_multiple_binary.R
\name{vec_ptype2.borg_sm_binary}
\alias{vec_ptype2.borg_sm_binary}
\title{Boiler plate for coercion for coercion of select multiple binarys}
\usage{
\method{vec_ptype2}{borg_sm_binary}(x, y, ...)
}
\description{
Boiler plate for coercion for coercion of select multiple binarys
}
