% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_select_multiple.R
\name{vec_ptype2.borg_sm_char}
\alias{vec_ptype2.borg_sm_char}
\title{Boiler plate for coercion for coercion of select multiple characters}
\usage{
\method{vec_ptype2}{borg_sm_char}(x, y, ...)
}
\description{
Boiler plate for coercion for coercion of select multiple characters
}
