% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_select_multiple.R
\name{vec_ptype2.borg_sm_list}
\alias{vec_ptype2.borg_sm_list}
\title{Boiler plate for coercion of select multiple lists}
\usage{
\method{vec_ptype2}{borg_sm_list}(x, y, ...)
}
\description{
Boiler plate for coercion of select multiple lists
}
