% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_text.R
\name{vec_ptype2.borg_text}
\alias{vec_ptype2.borg_text}
\title{Boiler plate for coercion for coercion of texts}
\usage{
\method{vec_ptype2}{borg_text}(x, y, ...)
}
\description{
Boiler plate for coercion for coercion of texts
}
