% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_select_multiple_helper.R
\name{vec_ptype2.char_helper}
\alias{vec_ptype2.char_helper}
\title{Boiler plate for coercion for coercion of texts}
\usage{
\method{vec_ptype2}{char_helper}(x, y, ...)
}
\description{
Boiler plate for coercion for coercion of texts
}
