% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_calc_character.R
\name{vec_ptype_abbr.borg_calc_character}
\alias{vec_ptype_abbr.borg_calc_character}
\title{Partial abbreviation in tibbles}
\usage{
\method{vec_ptype_abbr}{borg_calc_character}(x, ...)
}
\description{
Partial abbreviation in tibbles
}
