% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_calendar.R
\name{vec_ptype_abbr.borg_calendar}
\alias{vec_ptype_abbr.borg_calendar}
\title{Partial abbreviation in tibbles}
\usage{
\method{vec_ptype_abbr}{borg_calendar}(x, ...)
}
\description{
Partial abbreviation in tibbles
}
