% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_number.R
\name{vec_ptype_abbr.borg_number}
\alias{vec_ptype_abbr.borg_number}
\title{Partial abbreviation in tibbles}
\usage{
\method{vec_ptype_abbr}{borg_number}(x, ...)
}
\description{
Partial abbreviation in tibbles
}
