% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_select_one.R
\name{vec_ptype_abbr.borg_slct_one}
\alias{vec_ptype_abbr.borg_slct_one}
\title{Partial abbreviation in tibbles}
\usage{
\method{vec_ptype_abbr}{borg_slct_one}(x, ...)
}
\description{
Partial abbreviation in tibbles
}
