% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_select_multiple.R, R/class_select_one.R
\name{vec_ptype_abbr.borg_sm_char}
\alias{vec_ptype_abbr.borg_sm_char}
\title{Partial abbreviation in tibbles}
\usage{
\method{vec_ptype_abbr}{borg_sm_char}(x, ...)

\method{vec_ptype_abbr}{borg_sm_char}(x, ...)
}
\description{
Partial abbreviation in tibbles

Partial abbreviation in tibbles
}
