% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_select_multiple.R
\name{vec_ptype_abbr.borg_sm_list}
\alias{vec_ptype_abbr.borg_sm_list}
\title{Partial abbreviation in tibbles}
\usage{
\method{vec_ptype_abbr}{borg_sm_list}(x, ...)
}
\description{
Partial abbreviation in tibbles
}
