% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_calc_character.R
\name{vec_ptype_full.borg_calc_character}
\alias{vec_ptype_full.borg_calc_character}
\title{Full abbreviation in tibbles}
\usage{
\method{vec_ptype_full}{borg_calc_character}(x, ...)
}
\description{
Full abbreviation in tibbles
}
