% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_calendar.R
\name{vec_ptype_full.borg_calendar}
\alias{vec_ptype_full.borg_calendar}
\title{Full abbreviation in tibbles}
\usage{
\method{vec_ptype_full}{borg_calendar}(x, ...)
}
\description{
Full abbreviation in tibbles
}
