% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_rank.R
\name{vec_ptype_full.borg_rank_list}
\alias{vec_ptype_full.borg_rank_list}
\title{Full abbreviation in tibbles}
\usage{
\method{vec_ptype_full}{borg_rank_list}(x, ...)
}
\description{
Full abbreviation in tibbles
}
