% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_rank_position.R
\name{vec_ptype_full.borg_rank_position}
\alias{vec_ptype_full.borg_rank_position}
\title{Full abbreviation in tibbles}
\usage{
\method{vec_ptype_full}{borg_rank_position}(x, ...)
}
\description{
Full abbreviation in tibbles
}
