% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_select_one.R
\name{vec_ptype_full.borg_slct_one}
\alias{vec_ptype_full.borg_slct_one}
\title{Full abbreviation in tibbles}
\usage{
\method{vec_ptype_full}{borg_slct_one}(x, ...)
}
\description{
Full abbreviation in tibbles
}
