% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_select_multiple_binary.R
\name{vec_ptype_full.borg_sm_binary}
\alias{vec_ptype_full.borg_sm_binary}
\title{Full abbreviation in tibbles}
\usage{
\method{vec_ptype_full}{borg_sm_binary}(x, ...)
}
\description{
Full abbreviation in tibbles
}
