% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_select_multiple_helper.R
\name{vec_ptype_full.char_helper}
\alias{vec_ptype_full.char_helper}
\title{Full abbreviation in tibbles}
\usage{
\method{vec_ptype_full}{char_helper}(x, ...)
}
\description{
Full abbreviation in tibbles
}
