% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks.R
\name{attr_sm_err}
\alias{attr_sm_err}
\title{Check additional attributes for borg select multiple classes}
\usage{
attr_sm_err(binary_type, binary_sep)
}
\description{
Check additional attributes for borg select multiple classes
}
