% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks_rnk.R
\name{rnk_pos_sep_error}
\alias{rnk_pos_sep_error}
\title{Check binary separator attribute for borg select multiple classes}
\usage{
rnk_pos_sep_error(x)
}
\description{
Check binary separator attribute for borg select multiple classes
}
